mythtranscode: rename commandlineparser.* to mythtranscode_…
