+++ mythtv/programs/mythtranscode/mythtranscode_commandlineparser.h
